requirements.txt
streamlit
Pillow
imageio
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib